refactor(sequence): delete plugin and move to boot
