--- api-boot-project/api-boot-plugins/api-boot-plugin-sequence/pom.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>api-boot-plugins</artifactId>
-        <groupId>org.minbox.framework</groupId>
-        <version>${revision}</version>
-    </parent>
-    <artifactId>api-boot-plugin-sequence</artifactId>
-    <packaging>jar</packaging>
-    <description>
-        分布式高效ID算法集成，基于Twitter的Snowflake（"雪花算法"）算法实现分布式高效有序ID生产黑科技
-
-        Distributed efficient ID algorithm integration,
-        based on Twitter's Snowflake algorithm to achieve distributed efficient and orderly ID production black
-        technology.
-
-        See more details：https://gitee.com/yu120/sequence
-    </description>
-    <modelVersion>4.0.0</modelVersion>
-    <properties>
-        <main.basedir>${basedir}/../../..</main.basedir>
-    </properties>
-
-</project>
